requirements.txt
# === 🌐 Web App (Flask) ===
flask>=2.2.5
flask-cors
jinja2
werkzeug

# === 🔗 Base de données MongoDB ===
pymongo>=4.6.1

# === 📊 Data Science ===
pandas>=2.2.1
numpy
scipy
matplotlib==3.10.1
openpyxl>=3.1.2

# === 🤖 IA & NLP ===
scikit-learn>=1.4.1
transformers>=4.39.3
torch>=2.2.1
huggingface_hub
llama-cpp-python==0.3.8

# === 🌐 Scraping ===
requests
beautifulsoup4
lxml
yfinance==0.2.55

# === 🔒 Environnement & Sécurité ===
python-dotenv>=1.0.1

# === ⚙️ Automatisation ===
schedule==1.2.0
matplotlib-inline==0.1.7

# === 🧪 Tests ===
pytest

# === 🔄 Cache & Backend Utils ===
diskcache
